Deep_agent/requirements.txt
deepagents
langchain
langchain-core
langchain-openai
langchain-groq
ipykernel
tavily-python
python-dotenv